fix: move some unit as integration, mock vertexaiembeddings and apply ruff
