+++ libs/vertexai/tests/integration_tests/test_embeddings.py
@@ -4,6 +4,8 @@ Your end-user credentials would be used to make the calls (make sure you've run
 `gcloud auth login` first).
 """
 import pytest
+from vertexai.language_models import TextEmbeddingModel
+from vertexai.vision_models import MultiModalEmbeddingModel
 
 from langchain_google_vertexai.embeddings import VertexAIEmbeddings
 
@@ -80,3 +82,15 @@ def test_langchain_google_vertexai_image_embeddings(tmp_image) -> None:
     model = VertexAIEmbeddings(model_name="multimodalembedding")
     output = model.embed_image(tmp_image)
     assert len(output) == 1408
+
+
+def test_langchain_google_vertexai_text_model() -> None:
+    embeddings_model = VertexAIEmbeddings(model_name="textembedding-gecko@001")
+    assert isinstance(embeddings_model.client, TextEmbeddingModel)
+    assert not embeddings_model._is_multimodal_model(embeddings_model.model_name)
+
+
+def test_langchain_google_vertexai_multimodal_model() -> None:
+    embeddings_model = VertexAIEmbeddings(model_name="multimodalembedding@001")
+    assert isinstance(embeddings_model.client, MultiModalEmbeddingModel)
+    assert embeddings_model._is_multimodal_model(embeddings_model.model_name)
